+++ version_history.md
@@ -1,5 +1,10 @@
 #Version history#
 
+##0.2 (08/11/12)
+* Added Comments, syntax cleaning
+* Renamed method *getRelaxation()* to *getRelaxationParameter()*
+* Renamed method *setRelaxation()* to *setRelaxationParameter()*
+
 ##0.1 (05/29/12)
 * Initial release